requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-monitor-opentelemetry==1.6.4
azure-monitor-opentelemetry-exporter==1.0.0b32
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.15
exceptiongroup==1.2.2
fastapi==0.115.5
fixedint==0.1.6
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
msrest==0.7.1
oauthlib==3.2.2
opentelemetry-api==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-dbapi==0.49b1
opentelemetry-instrumentation-django==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-instrumentation-flask==0.49b1
opentelemetry-instrumentation-psycopg2==0.49b1
opentelemetry-instrumentation-requests==0.49b1
opentelemetry-instrumentation-urllib==0.49b1
opentelemetry-instrumentation-urllib3==0.49b1
opentelemetry-instrumentation-wsgi==0.49b1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
packaging==24.2
psutil==5.9.8
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
wrapt==1.16.0
zipp==3.21.0
